@@ -1,13 +1,12 @@
-﻿<Project ToolsVersion="15.0" Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <RuntimeIdentifiers>win7-x64;rhel.7-x64;ubuntu.16.04-x64</RuntimeIdentifiers>
-    <PreserveCompilationContext>true</PreserveCompilationContext>
     <CreateUser>true</CreateUser>
     <InstallService>true</InstallService>
-    <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
   </PropertyGroup>
+
   <ItemGroup>
     <Content Update="appsettings.json">
         <LinuxPath>/etc/demo/appsettings.json</LinuxPath>
@@ -21,11 +20,10 @@
     <RpmDependency Include="libunwind" Version="1.1" />
     <RpmDependency Include="libicu" Version="50.1.2" />
   </ItemGroup>
+
   <ItemGroup>
     <PackageReference Include="Nerdbank.GitVersioning" Version="2.3.38" PrivateAssets="all" />
     <PackageReference Include="Packaging.Targets" Version="$(PackagingNuGetVersion)" />
   </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.6" />
-  </ItemGroup>
+
 </Project>
